Post Muon TDR ME0 geometry cleanup
