+++ Geometry/GEMGeometry/interface/ME0EtaPartition.h
@@ -12,7 +12,7 @@ class ME0EtaPartitionSpecs;
 class ME0EtaPartition : public GeomDet
 {
 public:
-  
+
   ME0EtaPartition(ME0DetId id, const BoundPlane::BoundPlanePointer& bp, ME0EtaPartitionSpecs* rrs);
   ~ME0EtaPartition() override;
 
@@ -25,11 +25,11 @@ public:
   const Topology& padTopology() const;
   const StripTopology& specificPadTopology() const;
 
-  const GeomDetType& type() const override; 
- 
-  /// Return the chamber this roll belongs to 
+  const GeomDetType& type() const override;
+
+  /// Return the chamber this roll belongs to
   //const ME0Chamber* chamber() const;
- 
+
   // strip-related methods:
 
   /// number of readout strips in partition
@@ -42,7 +42,7 @@ public:
   /// returns center of strip position for FRACTIONAL strip number
   /// that has a value range of [0., nstrip]
   LocalPoint  centreOfStrip(float strip) const;
-  LocalError  localError(float strip) const;
+  LocalError  localError(float strip, float cluster_size = 1.) const;
 
   /// returns fractional strip number [0..nstrips] for a LocalPoint
   /// E.g., if local point hit strip #2, the fractional strip number would be
@@ -51,10 +51,10 @@ public:
 
   float pitch() const;
   float localPitch(const LocalPoint& lp) const;
- 
+
 
   // ME0-CSC pad-related methods:
-  
+
   /// number of ME0-CSC trigger readout pads in partition
   int npads() const;
 
@@ -76,7 +76,7 @@ public:
 
 
   // relations between strips and pads:
-  
+
   /// returns FRACTIONAL pad number [0.,npads] for an integer strip [1,nstrip]
   float padOfStrip(int strip) const;
