+++ Geometry/GEMGeometry/test/BuildFile.xml
@@ -5,8 +5,6 @@
 <flags   EDM_PLUGIN="1"/>
 <library   file="GEMGeometryAnalyzer.cc" name="GEMGeometryAnalyzer">
 </library>
-<library   file="ME0GeometryAnalyzer.cc" name="ME0GeometryAnalyzer1EtaPart">
-</library>
-<library   file="ME0GeometryAnalyzer10EtaPart.cc" name="ME0GeometryAnalyzer10EtaPart">
+<library   file="ME0GeometryAnalyzer.cc" name="ME0GeometryAnalyzer">
 </library>
