+++ Geometry/GEMGeometry/test/testGEMGeometry_cfg.py
@@ -2,13 +2,8 @@ import FWCore.ParameterSet.Config as cms
 
 process = cms.Process("Demo")
 
-
-process.load('Configuration.Geometry.GeometryExtended2023D4Reco_cff')
-process.load('Configuration.Geometry.GeometryExtended2023D4_cff')
-
-
-#process.load('Configuration.Geometry.GeometryExtended2023D1_cff')
-#process.load('Configuration.Geometry.GeometryExtended2023D1Reco_cff')
+process.load('Configuration.Geometry.GeometryExtended2023D21Reco_cff')
+process.load('Configuration.Geometry.GeometryExtended2023D21_cff')
 process.load('Configuration.StandardSequences.FrontierConditions_GlobalTag_cff')
 from Configuration.AlCa.GlobalTag import GlobalTag
 process.GlobalTag = GlobalTag(process.GlobalTag, 'auto:run2_mc', '')
@@ -37,10 +32,10 @@ process.p = cms.Path(process.test)
 
 ### TO ACTIVATE LogTrace NEED TO COMPILE IT WITH:
 ### -----------------------------------------------------------
-### --> scram b -j8 USER_CXXFLAGS="-DEDM_ML_DEBUG"             
-### Make sure that you first cleaned your CMSSW version:       
-### --> scram b clean                                          
-### before issuing the scram command above                     
+### --> scram b -j8 USER_CXXFLAGS="-DEDM_ML_DEBUG"
+### Make sure that you first cleaned your CMSSW version:
+### --> scram b clean
+### before issuing the scram command above
 ###############################################################
 process.load("FWCore.MessageLogger.MessageLogger_cfi")
 # process.MessageLogger.categories.append("GEMGeometryBuilderFromDDD")
@@ -53,4 +48,4 @@ process.MessageLogger.cout = cms.untracked.PSet(
     FwkReport = cms.untracked.PSet( limit = cms.untracked.int32(-1) ),
     # GEMGeometryBuilderFromDDD   = cms.untracked.PSet( limit = cms.untracked.int32(-1) ),
     # GEMNumberingScheme            = cms.untracked.PSet( limit = cms.untracked.int32(-1) ),
-)                                                                                                                                                                 
+)
