+++ Geometry/GEMGeometryBuilder/data/v7/GEMSpecs.xml
@@ -13,6 +13,10 @@
       <PartSelector path="//GHA2.."/>
       <Parameter name="nStrips" value="768"/>
     </SpecPar>
+    <SpecPar name="nPadsME0" eval="true">
+      <PartSelector path="//GHA0.."/>
+      <Parameter name="nPads" value="192"/>
+    </SpecPar>
     <SpecPar name="nPadsGE11" eval="true">
       <PartSelector path="//GHA1.."/>
       <Parameter name="nPads" value="192"/>
