+++ Geometry/GEMGeometryBuilder/plugins/ME0GeometryESModule.h
@@ -2,7 +2,7 @@
 #define Geometry_GEMGeometryBuilder_ME0GeometryESModule_h
 
 /** \class ME0GeometryESModule
- * 
+ *
  *  ESProducer for ME0Geometry in MuonGeometryRecord
  *
  *  \author M. Maggi - INFN Bari
@@ -15,7 +15,7 @@
 
 #include <memory>
 
-class ME0GeometryESModule : public edm::ESProducer 
+class ME0GeometryESModule : public edm::ESProducer
 {
  public:
   /// Constructor
@@ -23,12 +23,12 @@ class ME0GeometryESModule : public edm::ESProducer
 
   /// Destructor
   ~ME0GeometryESModule() override;
-  
+
   /// Produce ME0Geometry.
   std::shared_ptr<ME0Geometry>  produce(const MuonGeometryRecord & record);
-  
- private:  
+
+ private:
   // use the DDD as Geometry source
-  bool useDDD, use10EtaPart;
+  bool useDDD;
 };
 #endif
