+++ Geometry/GEMGeometryBuilder/src/ME0GeometryBuilderFromDDD.h
@@ -1,13 +1,7 @@
 #ifndef Geometry_GEMGeometry_ME0GeometryBuilderFromDDD_H
 #define Geometry_GEMGeometry_ME0GeometryBuilderFromDDD_H
 
-/** \class  ME0GeometryBuilderFromDDD
- *  Build the ME0Geometry ftom the DDD description
- *
- *  \author M. Maggi - INFN Bari
- *
- */
-
+#include "DataFormats/GeometrySurface/interface/Plane.h"
 #include <string>
 #include <map>
 #include <vector>
@@ -16,11 +10,13 @@ class DDCompactView;
 class DDFilteredView;
 class ME0Geometry;
 class ME0DetId;
+class ME0Chamber;
+class ME0Layer;
 class ME0EtaPartition;
 class MuonDDDConstants;
 
-class ME0GeometryBuilderFromDDD 
-{ 
+class ME0GeometryBuilderFromDDD
+{
  public:
 
   ME0GeometryBuilderFromDDD();
@@ -32,6 +28,17 @@ class ME0GeometryBuilderFromDDD
  private:
   ME0Geometry* buildGeometry(DDFilteredView& fview, const MuonDDDConstants& muonConstants);
   std::map<ME0DetId,std::vector<ME0DetId>> chids;
+
+  typedef ReferenceCountingPointer<BoundPlane> ME0BoundPlane;
+
+  ME0BoundPlane boundPlane(const DDFilteredView& fv,
+                           Bounds* bounds, bool isOddChamber) const ;
+
+  ME0Chamber* buildChamber(DDFilteredView& fv, ME0DetId detId) const;
+
+  ME0Layer* buildLayer(DDFilteredView& fv, ME0DetId detId) const;
+
+  ME0EtaPartition* buildEtaPartition(DDFilteredView& fv, ME0DetId detId) const;
 };
 
 #endif
